refactor questions page into components
